finrb [Refactor]: split Utils into financial domains
Extract accounting, ratio, return, yield, and TVM calculations into dedicated namespaces,
remove the duplicate Utils IRR implementation, and reorganize specs and API examples around
the new public API.
